pyproject.toml
[tool.poetry]
name = "mochi-api-client"
version = "0.1.1"
description = "Python wrapper for the Mochi spaced repetition platform API"
authors = ["GSejas <jsequeira03@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
pytest-cov = "^6.0.0"
pylint = "^3.3.1"
black = "^24.10.0"
allure-pytest = "^2.13.5"
safety = "^3.2.11"
interrogate = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
